Reworking how the data is sent to the front end. Changed blog modal to one to many relationship with comments. Comments no longer link to the blog it was posted to.

+++ controllers/commentController.js
@@ -1,48 +1,52 @@
 const Comment = require("../models/Comment");
+const Blog = require("../models/Blog");
+const { populate } = require("../models/Blog");
 
 exports.comment_get = (req, res, next) => {
   Comment.find().exec((err, comments) => {
     if (err) return next(err);
-    res.json({ current_user: req.user.email, comments });
+    res.json({ comments });
   });
 };
 
 exports.blog_comment_get = (req, res, next) => {
-  Comment.find({ $post: req.params.id }).exec(function (err, comments) {
-    if (err) {
-      return next(err);
-    }
-    res.json({ current_user: req.user.email, comments });
-  });
+  Blog.findById(req.params.id)
+    .populate({ path: "comments", populate: { path: "user", model: "User" } })
+    .exec(function (err, result) {
+      if (err) return next(err);
+      res.json({ result });
+    });
 };
 
 exports.blog_comment_post = (req, res, next) => {
-  if (req.user) {
-    const newComment = new Comment({
-      post: req.params.id,
-      user: req.user.id,
-      text: req.body.text,
-      public_username: "",
-    }).save((err, comment) => {
-      if (err) return next(err);
-      res.json({ current_user: req.user.email, message: "success", comment });
-    });
-  } else {
-    const newComment = new Comment({
-      post: req.params.id,
-      public_username: req.body.anonymous,
-      text: req.body.text,
-    }).save((err, comment) => {
-      if (err) return next(err);
-      res.json({ message: "success", comment });
-    });
-  }
+  const newComment = new Comment({
+    post: req.params.id,
+    user: req.user ? req.user.id : null,
+    text: req.body.text,
+    public_username: req.user
+      ? null
+      : req.body.anonymous
+      ? req.body.anonymous
+      : "Anonymous",
+  }).save((err, comment) => {
+    if (err) return next(err);
+    Blog.findByIdAndUpdate(
+      req.params.id,
+      {
+        $push: { comments: comment._id },
+      },
+      function (err, result) {
+        if (err) return next(err);
+        res.json({ result });
+      }
+    );
+  });
 };
 
 exports.comment_detail = (req, res, next) => {
   Comment.findById(req.params.id).exec(function (err, comment) {
     if (err) return next(err);
-    res.json({ current_user: req.user.email, comment });
+    res.json({ comment });
   });
 };
 
@@ -59,7 +63,7 @@ exports.comment_update_put = (req, res, next) => {
       { returnDocument: "after" },
       (err, result) => {
         if (err) return next(err);
-        res.json({ current_user: req.user.email, result });
+        res.json({ result });
       }
     );
   } else {
@@ -68,12 +72,15 @@ exports.comment_update_put = (req, res, next) => {
 };
 
 exports.comment_delete = (req, res, next) => {
-  if (req.user && req.user.admin) {
+  if (req.user && req.user.admin && req.body.blog) {
     Comment.findByIdAndDelete(req.params.id).exec(function (err, result) {
       if (err) return next(err);
-      res.json({
-        current_user: req.user.email,
-        message: `Deleted comment ${req.params.id}`,
+      Blog.updateOne(
+        { _id: req.body.blog },
+        { $pull: { comments: req.params.id } }
+      ).exec(function (err, doc) {
+        if (err) return next(err);
+        res.json({ doc });
       });
     });
   } else {
